server2/requirements.txt
django
pymongo == 2.8.1
mongoengine == 0.9.0
djangorestframework 